% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alquerque.r
\name{alquerque_games_variant}
\alias{alquerque_games_variant}
\alias{alquerque_alquerque}
\title{Setups for alquerque variants}
\usage{
alquerque_alquerque()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{tibble} data frames of setups for 1 alquerque variants.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   Alquerque \tab \url{https://en.wikipedia.org/wiki/Alquerque} \cr
}
}
\examples{
df <- alquerque_alquerque()
if (requireNamespace("ppcli", quietly = TRUE) &&
    packageVersion("ppcli") >= "0.2.0-1") {
  ppcli::cat_piece(df)
}
}
