% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker_games_other.r
\name{checker_games_other}
\alias{checker_games_other}
\alias{checker_breakthrough}
\alias{checker_crossings}
\alias{checker_dao}
\alias{checker_focus}
\alias{checker_four_field_kono}
\alias{checker_grasshopper}
\alias{checker_julgonu}
\alias{checker_lines_of_action}
\title{Setups for other games playable with a checkers set}
\usage{
checker_breakthrough(cell_width = 1)

checker_crossings(cell_width = 1)

checker_dao(cell_width = 1)

checker_focus(cell_width = 1)

checker_four_field_kono(cell_width = 1)

checker_grasshopper(cell_width = 1)

checker_julgonu(cell_width = 1)

checker_lines_of_action(cell_width = 1)
}
\arguments{
\item{cell_width}{Width of board cell.  Most renderers support \code{1} or \code{2}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{tibble} data frames of setups for 8 other games playable with a checkers set.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   Breakthrough \tab \url{https://en.wikipedia.org/wiki/Breakthrough_(board_game)} \cr
   Crossings \tab \url{https://en.wikipedia.org/wiki/Crossings_(game)} \cr
   Dao \tab \url{https://boardgamegeek.com/boardgame/948/dao} \cr
   Focus \tab \url{https://en.wikipedia.org/wiki/Focus_(board_game)} \cr
   Four Field Kono \tab \url{https://en.wikipedia.org/wiki/Four_Field_Kono} \cr
   Grasshopper \tab \url{http://www.cyningstan.com/game/71/grasshopper} \cr
   Jul-Gonu \tab \url{https://en.wikipedia.org/wiki/Jul-Gonu} \cr
   Lines of Action \tab \url{https://en.wikipedia.org/wiki/Lines_of_Action} \cr
}
}
