% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker_games_variant.r
\name{checker_games_variant}
\alias{checker_games_variant}
\alias{checker_american_checkers}
\alias{checker_american_pool_checkers}
\alias{checker_bashni}
\alias{checker_brazilian_checkers}
\alias{checker_canadian_checkers}
\alias{checker_checkers}
\alias{checker_column_checkers}
\alias{checker_corner_checkers}
\alias{checker_czech_checkers}
\alias{checker_dameo}
\alias{checker_english_checkers}
\alias{checker_frisian_checkers}
\alias{checker_gothic_checkers}
\alias{checker_international_checkers}
\alias{checker_italian_checkers}
\alias{checker_jamaican_checkers}
\alias{checker_lasca}
\alias{checker_malaysian_checkers}
\alias{checker_one_way_checkers}
\alias{checker_portuguese_checkers}
\alias{checker_russian_checkers}
\alias{checker_singaporean_checkers}
\alias{checker_spanish_checkers}
\alias{checker_thai_checkers}
\alias{checker_turkish_checkers}
\alias{checker_unified_pool_checkers}
\alias{checker_zimbabwean_pool_checkers}
\title{Setups for checkers variants}
\usage{
checker_american_checkers(cell_width = 1)

checker_american_pool_checkers(cell_width = 1)

checker_bashni(cell_width = 1)

checker_brazilian_checkers(cell_width = 1)

checker_canadian_checkers(cell_width = 1)

checker_checkers(cell_width = 1)

checker_column_checkers(cell_width = 1)

checker_corner_checkers(cell_width = 1)

checker_czech_checkers(cell_width = 1)

checker_dameo(cell_width = 1)

checker_english_checkers(cell_width = 1)

checker_frisian_checkers(cell_width = 1)

checker_gothic_checkers(cell_width = 1)

checker_international_checkers(cell_width = 1)

checker_italian_checkers(cell_width = 1)

checker_jamaican_checkers(cell_width = 1)

checker_lasca(cell_width = 1)

checker_malaysian_checkers(cell_width = 1)

checker_one_way_checkers(cell_width = 1)

checker_portuguese_checkers(cell_width = 1)

checker_russian_checkers(cell_width = 1)

checker_singaporean_checkers(cell_width = 1)

checker_spanish_checkers(cell_width = 1)

checker_thai_checkers(cell_width = 1)

checker_turkish_checkers(cell_width = 1)

checker_unified_pool_checkers(cell_width = 1)

checker_zimbabwean_pool_checkers(cell_width = 1)
}
\arguments{
\item{cell_width}{Width of board cell.  Most renderers support \code{1} or \code{2}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{tibble} data frames of setups for 23 checkers variants.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   (American) Checkers AKA (English) Draughts \tab \url{https://en.wikipedia.org/wiki/Draughts} \cr
   American Pool Checkers \tab \url{https://draughts.github.io/american-pool-checkers.html} \cr
   Bashni AKA Column Checkers \tab \url{https://en.wikipedia.org/wiki/Bashni} \cr
   Brazilian Checkers AKA Brazilian Draughts \tab \url{https://en.wikipedia.org/wiki/Brazilian_draughts} \cr
   Canadian Checkers AKA Canadian Draughts \tab \url{https://en.wikipedia.org/wiki/Canadian_checkers} \cr
   Corner Checkers AKA Corner Draughts \tab \url{https://brainking.com/en/GameRules?tp=30} \cr
   Czech Checkers AKA Czech Draughts \tab \url{https://en.wikipedia.org/wiki/Czech_draughts} \cr
   Dameo \tab \url{https://en.wikipedia.org/wiki/Dameo} \cr
   Frisian Checkers AKA Frisian Draughts \tab \url{https://en.wikipedia.org/wiki/Frisian_draughts} \cr
   Gothic Checkers AKA Gothic Draughts \tab \url{http://mlwi.magix.net/bg/gothiccheckersvariants.htm} \cr
   International Checkers AKA International Draughts \tab \url{https://en.wikipedia.org/wiki/International_draughts} \cr
   Italian Checkers AKA Italian Draughts \tab \url{https://en.wikipedia.org/wiki/Italian_draughts} \cr
   Jamaican Checkers AKA Jamaican Draughts \tab \url{https://web.archive.org/web/20230605023244/http://poolcheckers.com/jamaica/} \cr
   Lasca \tab \url{http://www.lasca.org} \cr
   Malaysian Checkers AKA Singaporean Checkers \tab \url{https://brainking.com/en/GameRules?tp=31} \cr
   One Way Checkers AKA One Way Draughts \tab \url{https://brainking.com/en/GameRules?tp=31} \cr
   Portuguese Checkers AKA Portuguese Draughts \tab \url{http://www.fpdamas.pt/regras/} \cr
   Russian Checkers AKA Russian Draughts \tab \url{https://en.wikipedia.org/wiki/Russian_draughts} \cr
   Spanish Checkers AKA Spanish Draughts \tab \url{https://mindsports.nl/index.php/on-the-evolution-of-draughts-variants/draughts-variants/497-dama_s} \cr
   Thai Checkers AKA Thai Draughts AKA Mak-hot AKA Makhos \tab \url{https://checkers.fandom.com/wiki/Mak-hot} \cr
   Turkish Checkers AKA Turkish Draughts AKA Dama \tab \url{https://en.wikipedia.org/wiki/Turkish_draughts} \cr
   Unified Pool Checkers \tab \url{https://wiegerw.github.io/pdn/gametype.html} \cr
   Zimbabwean Pool Checkers \tab \url{https://wiegerw.github.io/pdn/gametype.html} \cr
}
}
