% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker_boards.r
\name{checker_pieces}
\alias{checker_pieces}
\alias{checker_board}
\alias{checker_bits}
\title{Generate checkers boards and pieces}
\usage{
checker_board(
  nrows = 8L,
  ncols = nrows,
  x0 = 1,
  y0 = 1,
  ...,
  side = "face",
  piece_side = paste0("board_", side),
  suit = 3L,
  angle = 0,
  cell_width = 1
)

checker_bits(
  ...,
  suit = 1:6,
  x = as.double(1:6),
  y = 1,
  angle = 0,
  cell_width = 1,
  length.out = NA_integer_
)
}
\arguments{
\item{nrows}{Number of rows in game board}

\item{ncols}{Number of columns in game board}

\item{x0}{X coordinate for the center of the first cell}

\item{y0}{Y coordinate for the center of the first cell}

\item{...}{Should be left empty.}

\item{side}{Either "face" or "back"}

\item{piece_side}{If \code{"board_face"} a checkered board (as in American Checkers).
If \code{"board_back"} a lined board (as in Turkish Checkers).}

\item{suit}{Suit value (color) of board.
\code{1L} is "red", \code{2L} is "black", \code{3L} is "green", \code{4L} is "blue", \code{5L} is "yellow", and \code{6L} is "white".}

\item{angle}{Angle of board in degrees.
Italian Checkers rotates its board 90 degrees compared to other checkers variants.}

\item{cell_width}{Width of board cell.
\code{\link[piecepackr:game_systems]{piecepackr::game_systems()}} supports \code{1} or \code{2}.
If \code{NULL} then don't set it yet.}

\item{x, y}{Cartesian coordinates (numeric vectors)}

\item{length.out}{The number of pieces.
Not needed if all the arguments are the same length (or of length one)
and this length is the same as the number of desired pieces.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{checker_board()} returns a square checkers board of desired size and type.
\code{checker_bits()} returns checkers discs.
}
