% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers_games_other.r
\name{checkers_games_other}
\alias{checkers_games_other}
\alias{checkers_breakthrough}
\alias{checkers_crossings}
\alias{checkers_focus}
\alias{checkers_four_field_kono}
\alias{checkers_grasshopper}
\alias{checkers_julgonu}
\alias{checkers_lines_of_action}
\title{Setups for other games playable with a checkers set}
\usage{
checkers_breakthrough(cell_width = 1)

checkers_crossings(cell_width = 1)

checkers_focus(cell_width = 1)

checkers_four_field_kono(cell_width = 1)

checkers_grasshopper(cell_width = 1)

checkers_julgonu(cell_width = 1)

checkers_lines_of_action(cell_width = 1)
}
\arguments{
\item{cell_width}{Width of board cell.  Most renderers support \code{1} or \code{2}.}
}
\description{
\code{tibble} data frames of setups for 7 other games playable with a checkers set.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   Breakthrough \tab \url{https://en.wikipedia.org/wiki/Breakthrough_(board_game)} \cr
   Crossings \tab \url{https://en.wikipedia.org/wiki/Crossings_(game)} \cr
   Focus \tab \url{https://en.wikipedia.org/wiki/Focus_(board_game)} \cr
   Four Field Kono \tab \url{https://en.wikipedia.org/wiki/Four_Field_Kono} \cr
   Grasshopper \tab \url{http://www.cyningstan.com/game/71/grasshopper} \cr
   Jul-Gonu \tab \url{https://en.wikipedia.org/wiki/Jul-Gonu} \cr
   Lines of Action \tab \url{https://en.wikipedia.org/wiki/Lines_of_Action} \cr
}
}
