% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers_games_variant.r
\name{checkers_games_variant}
\alias{checkers_games_variant}
\alias{checkers_american_checkers}
\alias{checkers_brazilian_checkers}
\alias{checkers_canadian_checkers}
\alias{checkers_checkers}
\alias{checkers_czech_checkers}
\alias{checkers_dameo}
\alias{checkers_english_checkers}
\alias{checkers_frisian_checkers}
\alias{checkers_international_checkers}
\alias{checkers_italian_checkers}
\alias{checkers_portuguese_checkers}
\alias{checkers_russian_checkers}
\alias{checkers_thai_checkers}
\alias{checkers_turkish_checkers}
\title{Setups for checkers variants}
\usage{
checkers_american_checkers(cell_width = 1)

checkers_brazilian_checkers(cell_width = 1)

checkers_canadian_checkers(cell_width = 1)

checkers_checkers(cell_width = 1)

checkers_czech_checkers(cell_width = 1)

checkers_dameo(cell_width = 1)

checkers_english_checkers(cell_width = 1)

checkers_frisian_checkers(cell_width = 1)

checkers_international_checkers(cell_width = 1)

checkers_italian_checkers(cell_width = 1)

checkers_portuguese_checkers(cell_width = 1)

checkers_russian_checkers(cell_width = 1)

checkers_thai_checkers(cell_width = 1)

checkers_turkish_checkers(cell_width = 1)
}
\arguments{
\item{cell_width}{Width of board cell.  Most renderers support \code{1} or \code{2}.}
}
\description{
\code{tibble} data frames of setups for 12 checkers variants.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   Brazilian Checkers AKA Brazilian Draughts \tab \url{https://en.wikipedia.org/wiki/Brazilian_draughts} \cr
   Canadian Checkers AKA Canadian Draughts \tab \url{https://en.wikipedia.org/wiki/Canadian_checkers} \cr
   Czech Checkers AKA Czech Draughts \tab \url{https://en.wikipedia.org/wiki/Czech_draughts} \cr
   (American) Checkers AKA (English) Draughts \tab \url{https://en.wikipedia.org/wiki/Draughts} \cr
   Dameo \tab \url{https://en.wikipedia.org/wiki/Dameo} \cr
   Frisian Checkers AKA Frisian Draughts \tab \url{https://en.wikipedia.org/wiki/Frisian_draughts} \cr
   International Checkers AKA International Draughts \tab \url{https://en.wikipedia.org/wiki/International_draughts} \cr
   Italian Checkers AKA Italian Draughts \tab \url{https://en.wikipedia.org/wiki/Italian_draughts} \cr
   Portuguese Checkers AKA Portuguese Draughts \tab \url{http://www.fpdamas.pt/regras/} \cr
   Russian Checkers AKA Russian Draughts \tab \url{https://en.wikipedia.org/wiki/Russian_draughts} \cr
   Thai Checkers AKA Thai Draughts AKA Mak-hot AKA Makhos \tab \url{https://checkers.fandom.com/wiki/Mak-hot} \cr
   Turkish Checkers AKA Turkish Draughts AKA Dama \tab \url{https://en.wikipedia.org/wiki/Turkish_draughts} \cr
}
}
