% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_boards.r
\name{chess_board}
\alias{chess_board}
\title{Chess boards}
\usage{
chess_board(
  nrows = 8L,
  ncols = nrows,
  x0 = 1,
  y0 = 1,
  ...,
  piece_side = "board_face",
  suit = 3L,
  angle = 0,
  cell_width = NULL
)
}
\arguments{
\item{nrows}{Number of rows in game board}

\item{ncols}{Number of columns in game board}

\item{x0}{X coordinate for the center of the first cell}

\item{y0}{Y coordinate for the center of the first cell}

\item{...}{Ignored}

\item{piece_side}{If \code{"board_face"} a checkered board (as in American Checkers).
If \code{"board_back"} a lined board (as in Turkish Checkers).}

\item{suit}{Suit value (color) of board.  \code{2L} is "black" and \code{3L} is "green".}

\item{angle}{Angle of board in degrees.
Italian Checkers rotates its board 90 degrees compared to other checkers variants.}

\item{cell_width}{Width of board cell.
\code{\link[piecepackr:game_systems]{piecepackr::game_systems()}} supports \code{1} or \code{2}.
If \code{NULL} then don't set it yet.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{chess_board()} returns a square checkered board of desired size and type.
}
