% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_games_variant.R
\name{chess_games_variant}
\alias{chess_games_variant}
\alias{chess_chess}
\alias{chess_chess960}
\alias{chess_fischer_random_chess}
\alias{chess_horde_chess}
\alias{chess_international_chess}
\alias{chess_jeson_mor}
\alias{chess_monochrome_chess}
\alias{chess_racing_kings}
\title{Setups for chess variants}
\usage{
chess_chess(
  cell_width = getOption("ppdf.chess_cell_width", 1),
  ...,
  fen = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1"
)

chess_chess960(
  seed = NULL,
  cell_width = getOption("ppdf.chess_cell_width", 1),
  ...,
  fen = NULL
)

chess_fischer_random_chess(
  seed = NULL,
  cell_width = getOption("ppdf.chess_cell_width", 1),
  ...,
  fen = NULL
)

chess_horde_chess(
  ...,
  cell_width = getOption("ppdf.chess_cell_width", 1),
  fen = "rnbqkbnr/pppppppp/8/1PP2PP1/PPPPPPPP/PPPPPPPP/PPPPPPPP/PPPPPPPP w kq - 0 1"
)

chess_international_chess(
  cell_width = getOption("ppdf.chess_cell_width", 1),
  ...,
  fen = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1"
)

chess_jeson_mor(
  cell_width = getOption("ppdf.chess_cell_width", 1),
  ...,
  fen = "nnnnnnnnn/9/9/9/9/9/9/9/NNNNNNNNN"
)

chess_monochrome_chess(
  cell_width = getOption("ppdf.chess_cell_width", 1),
  ...,
  fen = "RNBQKBNR/PPPPPPPP/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1"
)

chess_racing_kings(
  cell_width = getOption("ppdf.chess_cell_width", 1),
  ...,
  fen = "8/8/8/8/8/8/krbnNBRK/qrbnNBRQ w - - 0 1"
)
}
\arguments{
\item{cell_width}{Width of board cell.  Most renderers support \code{1} or \code{2}.}

\item{...}{Should be left empty.}

\item{fen}{A FEN string specifying the piece placement.
Only the piece placement portion (the first field) of the FEN string is used.}

\item{seed}{Seed that determines setup, either an integer or \code{NULL}}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{tibble} data frames of setups for 6 chess variants.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   (International) Chess \tab \url{https://en.wikipedia.org/wiki/Chess} \cr
   Fischer Random Chess AKA Chess960 \tab \url{https://www.chessvariants.com/diffsetup.dir/fischer.html} \cr
   Horde Chess \tab \url{https://lichess.org/variant/horde} \cr
   Jeson Mor \tab \url{https://en.wikipedia.org/wiki/Jeson_Mor} \cr
   Monochrome Chess \tab \url{https://wunderland.com/WTS/Andy/Games/monochess.html} \cr
   Racing Kings \tab \url{https://www.chessvariants.com/diffobjective.dir/racing.html} \cr
}
}
\examples{
df <- chess_international_chess()
if (requireNamespace("ppcli", quietly = TRUE) &&
    packageVersion("ppcli") >= "0.3.0-1") {
  ppcli::cat_piece(df, annotate = TRUE)
}
}
