% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domino_games_variant.R
\name{domino_games_variant}
\alias{domino_games_variant}
\alias{domino_concentration}
\alias{domino_bee_donimoes}
\alias{domino_finder}
\alias{domino_freecell}
\alias{domino_fujisan}
\alias{domino_luzon}
\alias{domino_patience}
\alias{domino_runners}
\alias{domino_the_jubilee}
\alias{domino_the_sultan}
\title{Setups for dominoes variants}
\usage{
domino_concentration(seed = NULL)

domino_bee_donimoes(n = 7, seed = NULL)

domino_finder(seed = NULL)

domino_freecell(n = 7, seed = NULL)

domino_fujisan(seed = NULL, ..., pips = NULL, pawns = "S12M/A12C")

domino_luzon(seed = NULL)

domino_patience(seed = NULL)

domino_runners(seed = NULL)

domino_the_jubilee(seed = NULL)

domino_the_sultan(seed = NULL)
}
\arguments{
\item{seed}{Seed that determines setup, either an integer or \code{NULL}}

\item{n}{The number of doubles in a set e.g. \code{n = 7} for double-6 set,
\code{n = 10} for a double-9 set, \code{n = 13} for a double-12 set.}

\item{...}{Should be left empty.}

\item{pips}{A vector or matrix of Fuji-san pip layout, or \code{NULL} for a random layout.
Can also be a string of integers with the rows separated by a \code{/}.
The first row is the top row (y=2) and the second is the bottom row (y=1).}

\item{pawns}{A FEN-like string of the initial pawn placement.
The four pawns \code{S}, \code{M}, \code{C}, and \code{A} are represented by their letter and empty spaces by a number.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{tibble} data frames of setups for 10 dominoes variants.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   Bee Donimoes \tab \url{https://donkirkby.github.io/donimoes/rules.html#bee-donimoes} \cr
   Concentration \tab \url{http://www.domino-play.com/Games/Concentration.htm} \cr
   Domino Finder \tab \url{https://donkirkby.github.io/donimoes/rules.html#domino-finder} \cr
   Dominoes Freecell \tab \url{https://www.pagat.com/invented/domfreecell.html} \cr
   (Domino) Fuji-san \tab \url{http://donkirkby.github.io/donimoes/rules.html#fujisan} \cr
   Luzon \tab \url{http://www.domino-play.com/Games/Luzon.htm} \cr
   (Domino) Patience \tab \url{http://www.domino-play.com/Games/Patience.htm} \cr
   Domino Runners \tab \url{https://donkirkby.github.io/donimoes/rules.html#domino-runners} \cr
   The Jubilee \tab \url{http://www.domino-play.com/Games/Jubilee.htm} \cr
   The Sultan \tab \url{http://www.domino-play.com/Games/Sultan.htm} \cr
}
}
\examples{
df <- domino_the_jubilee()
if (require("piecepackr", quietly = TRUE) &&
    packageVersion("piecepackr") >= "1.15.0-9") {
  grid::grid.newpage()
  envir = game_systems(round = TRUE)
  pmap_piece(df, envir = envir, default.units = "in",
             trans = op_transform, op_scale = 0.5)
}
}
