% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domino_games_variant.r
\name{domino_tiles}
\alias{domino_tiles}
\title{Generate dominoes tiles}
\usage{
domino_tiles(
  n = 7,
  ...,
  side = "face",
  piece_side = paste0("tile_", side),
  suit = sequence(n:1, from = 1:n),
  rank = rep.int(1:n, n:1),
  cfg = "dominoes",
  x = sequence(n:1, from = 1:n),
  y = 2 * rep.int(n:1, n:1) - 0.5,
  angle = 0
)
}
\arguments{
\item{n}{The number of doubles in a set e.g. \code{n = 7} for double-6 set, \code{n = 10} for a double-9 set, \code{n = 13} for a double-12 set.}

\item{...}{Should be left empty.}

\item{side}{Either "face" or "back".}

\item{piece_side}{Either "tile_face" or "tile_back".}

\item{suit}{"Bottom" half of domino.  Will be coerced by \code{\link[=domino_suit]{domino_suit()}}.}

\item{rank}{"Top" half of domino.  Will be coerced by \code{\link[=piece_rank]{piece_rank()}}.}

\item{cfg}{"dominoes" or perhaps "dominoes_black", "dominoes_blue", "dominoes_green", "dominoes_red", "dominoes_white", or "dominoes_yellow".}

\item{x, y}{Cartesian coordinates (numeric vectors)}

\item{angle}{Rotation of dominoes (numeric vector of degrees, counter-clockwise).  Will be coerced by \code{\link[=piece_angle]{piece_angle()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{domino_tiles()} generates a data frame of dominoes tiles.
By default will generate all the dominoes tiles of a double-6 set.
}
\examples{
df_double6 <- domino_tiles()
nrow(df_double6)
df_double9 <- domino_tiles(n = 10)
nrow(df_double9)
df_double12 <- domino_tiles(n = 13)
nrow(df_double12)
df_double15 <- domino_tiles(n = 16)
nrow(df_double15)
}
