% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominoes_games_variant.r
\name{dominoes_games_variant}
\alias{dominoes_games_variant}
\alias{dominoes_fujisan}
\title{Setups for dominoes variants}
\usage{
dominoes_fujisan(seed = NULL)
}
\arguments{
\item{seed}{Seed that determines setup, either an integer or \code{NULL}}
}
\description{
\code{tibble} data frames of setups for 1 dominoes variants.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   (Domino) Fuji-san \tab \url{http://donkirkby.github.io/donimoes/rules.html#fujisan} \cr
}
}
