% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_info.r
\name{games_info}
\alias{games_info}
\alias{games_chess}
\alias{games_checkers}
\alias{games_dominoes}
\alias{games_piecepack}
\title{Data frames with game info}
\usage{
games_chess()

games_checkers()

games_dominoes()

games_piecepack()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with character columns "game", "methods", "comment", and "url"
for game name, setup function name(s), possible comment, and url for more information.
}
\description{
\code{games_chess()}, \code{games_dominoes()}, \code{games_checkers()}, and \code{game_piecepack()}
contain information about the games
whose setups are provided by this package.
}
\examples{
head(games_piecepack())
# Number of games for each game system
nrow(games_chess())
nrow(games_checkers())
nrow(games_dominoes())
nrow(games_piecepack())
}
