% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go_games_variant}
\alias{go_games_variant}
\alias{go_connect6}
\alias{go_go}
\alias{go_gomoku}
\alias{go_renju}
\title{Setups for go variants}
\usage{
go_connect6(nrows = 19L, ncols = nrows)

go_go(nrows = 19L, ncols = nrows)

go_gomoku(nrows = 15L, ncols = nrows)

go_renju(nrows = 15L, ncols = nrows)
}
\arguments{
\item{nrows}{Number of rows (lines) in game board}

\item{ncols}{Number of columns (lines) in game board}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{tibble} data frames of setups for 4 go variants.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   Connect6 \tab \url{https://en.wikipedia.org/wiki/Connect6} \cr
   Go \tab \url{https://en.wikipedia.org/wiki/Go_(game)} \cr
   Gomoku \tab \url{https://en.wikipedia.org/wiki/Gomoku} \cr
   Renju \tab \url{https://en.wikipedia.org/wiki/Renju} \cr
}
}
\examples{
df <- go_go()
if (requireNamespace("ppcli", quietly = TRUE) &&
    packageVersion("ppcli") >= "0.2.0-1") {
  ppcli::cat_piece(df, annotate = TRUE)
}
}
