% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris.r
\name{morris_games_variant}
\alias{morris_games_variant}
\alias{morris_three_mens_morris}
\alias{morris_five_mens_morris}
\alias{morris_six_mens_morris}
\alias{morris_seven_mens_morris}
\alias{morris_nine_mens_morris}
\alias{morris_ten_mens_morris}
\alias{morris_lasker_morris}
\alias{morris_twelve_mens_morris}
\title{Setups for morris variants}
\usage{
morris_three_mens_morris()

morris_five_mens_morris()

morris_six_mens_morris()

morris_seven_mens_morris()

morris_nine_mens_morris()

morris_ten_mens_morris()

morris_lasker_morris()

morris_twelve_mens_morris()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{tibble} data frames of setups for 7 morris variants.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   Three Men's Morris \tab \url{https://en.wikipedia.org/wiki/Three_men\%27s_morris} \cr
   Five Men's Morris \tab \url{https://en.wikipedia.org/wiki/Nine_men\%27s_morris#Six_men's_morris} \cr
   Six Men's Morris \tab \url{https://boardgamegeek.com/boardgame/25702/six-mens-morris} \cr
   Seven Men's Morris \tab \url{https://en.wikipedia.org/wiki/Nine_men\%27s_morris#Six_men's_morris} \cr
   Nine Men's Morris \tab \url{https://en.wikipedia.org/wiki/Nine_men\%27s_morris} \cr
   Ten Men's Morris aka Lasker Morris \tab \url{https://en.wikipedia.org/wiki/Nine_men\%27s_morris#Lasker_morris} \cr
   Twelve Men's Morris \tab \url{https://en.wikipedia.org/wiki/Morabaraba} \cr
}
}
\examples{
df <- morris_three_mens_morris()
if (requireNamespace("piecepackr", quietly = TRUE)) {
  piecepackr::render_piece(df, open_device = FALSE)
}
}
