% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris.r
\name{morris_pieces}
\alias{morris_pieces}
\alias{morris_board}
\alias{morris_bits}
\title{Generate morris boards and pieces}
\usage{
morris_board(n = 9L, x0 = 1, y0 = 1, ..., suit = "green")

morris_bits(
  ...,
  suit = 1:6,
  x = as.double(1:6),
  y = 1,
  angle = 0,
  length.out = NA_integer_
)
}
\arguments{
\item{n}{The number of "men" the morris board is for.}

\item{x0}{X coordinate for the center of the first point}

\item{y0}{Y coordinate for the center of the first point}

\item{...}{Should be left empty.}

\item{suit}{Suit value (color) of board/bit.
\code{1L} is "red", \code{2L} is "black", \code{3L} is "green", \code{4L} is "blue", \code{5L} is "yellow", and \code{6L} is "white".  Will be coerced by \code{\link[=morris_suit]{morris_suit()}}.}

\item{x, y}{Cartesian coordinates (numeric vectors)}

\item{angle}{Rotation of piece (numeric vector of degrees, counter-clockwise).
Will be coerced by \code{\link[=piecepack_angle]{piecepack_angle()}}.}

\item{length.out}{The number of pieces.
Not needed if all the arguments are the same length (or of length one)
and this length is the same as the number of desired pieces.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{morris_board()} returns various morris boards.
\code{morris_bits()} returns morris pieces aka merels.
}
