% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.r
\name{normalize_name}
\alias{normalize_name}
\title{Normalize name(s)}
\usage{
normalize_name(x, sep = "_")
}
\arguments{
\item{x}{Name to normalize.}

\item{sep}{Separator character.}
}
\value{
Character vector of normalized names.
}
\description{
\code{normalize_name()} normalizes (game) name(s).
}
\details{
\verb{normalize_name(x, sep = "_") is used to help create R variable names while }normalize_name(x, sep = "-")` is used to help create filenames and LaTeX labels.
}
\examples{
  normalize_name("Fuji-san")
  normalize_name("Nine Men's Morris", sep = "-")
}
