% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suits.r
\name{piece_suit}
\alias{piece_suit}
\alias{domino_suit}
\alias{tarot_suit}
\title{Cast to suit integers}
\usage{
piece_suit(suit)

domino_suit(suit)

tarot_suit(suit)
}
\arguments{
\item{suit}{A numeric or character vector of suit values.}
}
\value{
An integer vector of suit values.
}
\description{
These functions help make sure suit values are the
right integer values expected by the configurations used by
\code{\link[piecepackr:game_systems]{piecepackr::game_systems()}}  and \code{\link[ppn:read_ppn]{ppn::read_ppn()}}.
If the input is numeric it simply uses \code{\link[=as.integer]{as.integer()}} and if the
input is character it tries to cast to a reasonable value.
}
\examples{
piece_suit(c("red", "black", "green", "blue", "yellow", "white"))
piece_suit(c("R", "K", "G", "B", "Y", "W"))
piece_suit(c("suns", "moons", "crowns", "arms"))
piece_suit(c("S", "M", "C", "A"))
domino_suit(c("\U0001f06e", "\U0001f038"))
domino_suit(c("zero", "one", "eighteen"))
tarot_suit(c("\u2665", "\u2660", "\u2663", "\u2666"))
tarot_suit(c("hearts", "spades", "clubs", "diamonds", "trumps"))
tarot_suit(c("H", "S", "C", "D"))
}
