% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecepack_games_chess.r
\name{piecepack_games_chess}
\alias{piecepack_games_chess}
\alias{piecepack_alice_chess}
\alias{piecepack_baroque_chess}
\alias{piecepack_chaturaji}
\alias{piecepack_chess}
\alias{piecepack_chinese_chess}
\alias{piecepack_four_seasons_chess}
\alias{piecepack_international_chess}
\alias{piecepack_japanese_chess}
\alias{piecepack_minishogi}
\alias{piecepack_shogi}
\alias{piecepack_ultima}
\alias{piecepack_xiangqi}
\title{Setups for chess variants playable with a piecepack}
\usage{
piecepack_alice_chess(has_subpack = FALSE, max_tiles = 24L)

piecepack_baroque_chess(has_subpack = FALSE)

piecepack_chaturaji(has_subpack = FALSE)

piecepack_chess(has_subpack = FALSE)

piecepack_chinese_chess(has_subpack = FALSE)

piecepack_four_seasons_chess(has_subpack = FALSE)

piecepack_international_chess(has_subpack = FALSE)

piecepack_japanese_chess(has_subpack = FALSE, cfg2 = "piecepack")

piecepack_minishogi()

piecepack_shogi(has_subpack = FALSE, cfg2 = "piecepack")

piecepack_ultima(has_subpack = FALSE)

piecepack_xiangqi(has_subpack = FALSE)
}
\arguments{
\item{has_subpack}{Has a piecepack subpack}

\item{max_tiles}{Maximum number of (piecepack) tiles available to build boards}

\item{cfg2}{A string of a piecepack expansion (or perhaps \code{"piecepack"} for a second piecepack)}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{tibble} data frames of setups for 8 chess variants playable with a piecepack.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various chess variants:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   Alice Chess \tab \url{https://en.wikipedia.org/wiki/Alice_chess} \cr
   Chaturaji \tab \url{https://www.ludism.org/ppwiki/Chaturaji} \cr
   Four Seasons Chess \tab \url{https://www.chessvariants.com/historic.dir/4seiz.html} \cr
   (International) Chess \tab \url{https://www.ludism.org/ppwiki/Chess} \cr
   Minishogi \tab \url{https://en.wikipedia.org/wiki/Minishogi} \cr
   Shogi AKA Japanese Chess \tab \url{https://www.ludism.org/ppwiki/Shogi} \cr
   Ultima AKA Baroque Chess \tab \url{https://en.wikipedia.org/wiki/Baroque_chess} \cr
   Xiangqi AKA Chinese Chess \tab \url{https://www.ludism.org/ppwiki/Xiangqi} \cr
}
}
