% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecepack_games_original.r
\name{piecepack_games_original}
\alias{piecepack_games_original}
\alias{piecepack_alien_city}
\alias{piecepack_black_pawn_trucking}
\alias{piecepack_brain_burn}
\alias{piecepack_burbuja}
\alias{piecepack_cardinals_guards}
\alias{piecepack_cell_management}
\alias{piecepack_chariots}
\alias{piecepack_chinese_checkers}
\alias{piecepack_piecepack_halma}
\alias{piecepack_coin_collectors}
\alias{piecepack_climbing_man}
\alias{piecepack_crocodile_hop}
\alias{piecepack_desfases}
\alias{piecepack_easy_slider}
\alias{piecepack_everest}
\alias{piecepack_four_blind_mice}
\alias{piecepack_froggy_bottom}
\alias{piecepack_fujisan}
\alias{piecepack_galaxy_express}
\alias{piecepack_iceberg}
\alias{piecepack_ice_floe}
\alias{piecepack_japan}
\alias{piecepack_lab_rats}
\alias{piecepack_landlocked}
\alias{piecepack_ley_lines}
\alias{piecepack_mathrix}
\alias{piecepack_pawns_crossing}
\alias{piecepack_piecepackman}
\alias{piecepack_one_man_thrag}
\alias{piecepack_pass_the_food}
\alias{piecepack_piecepack_klondike}
\alias{piecepack_piece_gaps}
\alias{piecepack_piece_packing_pirates}
\alias{piecepack_plans_of_action}
\alias{piecepack_relativity}
\alias{piecepack_san_andreas}
\alias{piecepack_sarcophagus}
\alias{piecepack_shopping_mall}
\alias{piecepack_skyscrapers}
\alias{piecepack_slides_of_action}
\alias{piecepack_speedy_towers}
\alias{piecepack_steppin_stones}
\alias{piecepack_the_in_crowd}
\alias{piecepack_the_magic_bag}
\alias{piecepack_the_penguin_game}
\alias{piecepack_tower_of_babel}
\alias{piecepack_piecepack_accordion}
\alias{piecepack_tracers}
\alias{piecepack_triactor}
\alias{piecepack_tula}
\alias{piecepack_wormholes}
\title{Setups for various piecepack games}
\usage{
piecepack_alien_city(seed = NULL, tiles = NULL)

piecepack_black_pawn_trucking(seed = NULL)

piecepack_brain_burn(seed = NULL)

piecepack_burbuja(seed = NULL, tiles = NULL)

piecepack_cardinals_guards(seed = NULL, tiles = NULL)

piecepack_cell_management(seed = NULL)

piecepack_chariots()

piecepack_chinese_checkers()

piecepack_piecepack_halma()

piecepack_coin_collectors(seed = NULL, tiles = NULL, coins = NULL, dice = NULL)

piecepack_climbing_man(seed = NULL, variant = c("Basic", "Free"))

piecepack_crocodile_hop(seed = NULL)

piecepack_desfases(seed = NULL, tiles = NULL, dice = NULL)

piecepack_easy_slider(seed = NULL, tiles = NULL, coins = NULL, pawns = NULL)

piecepack_everest()

piecepack_four_blind_mice()

piecepack_froggy_bottom()

piecepack_fujisan(seed = NULL, coins = NULL, dice = NULL)

piecepack_galaxy_express(seed = NULL)

piecepack_iceberg(seed = NULL, tiles = NULL)

piecepack_ice_floe()

piecepack_japan(seed = NULL)

piecepack_lab_rats(seed = NULL)

piecepack_landlocked(seed = NULL)

piecepack_ley_lines()

piecepack_mathrix(seed = NULL, coins = NULL)

piecepack_pawns_crossing(..., seed = NULL, n_players = 2L)

piecepack_piecepackman(seed = NULL, variant = "Roundabout")

piecepack_one_man_thrag(seed = NULL)

piecepack_pass_the_food()

piecepack_piecepack_klondike(seed = NULL)

piecepack_piece_gaps(seed = NULL)

piecepack_piece_packing_pirates(seed = NULL)

piecepack_plans_of_action(seed = NULL, coins = NULL)

piecepack_relativity(seed = NULL, coins = NULL)

piecepack_san_andreas()

piecepack_sarcophagus(seed = NULL)

piecepack_shopping_mall(seed = NULL, cfg2 = "go")

piecepack_skyscrapers(seed = NULL, tiles = NULL)

piecepack_slides_of_action()

piecepack_speedy_towers(n_players = 2, seed = NULL)

piecepack_steppin_stones(seed = NULL)

piecepack_the_in_crowd()

piecepack_the_magic_bag(seed = NULL)

piecepack_the_penguin_game(seed = NULL)

piecepack_tower_of_babel(seed = NULL, tiles = NULL)

piecepack_piecepack_accordion(seed = NULL, tiles = NULL)

piecepack_tracers()

piecepack_triactor(seed = NULL, cfg2 = "playing_cards_expansion")

piecepack_tula(
  seed = NULL,
  tiles = NULL,
  variant = c("Original", "Variant 1", "Variant 2", "Variant 3", "Variant 4")
)

piecepack_wormholes()
}
\arguments{
\item{seed}{Seed that determines setup, either an integer or \code{NULL}}

\item{tiles}{String of tile layout}

\item{coins}{String of coin layout}

\item{dice}{String of dice layout}

\item{variant}{Name of variant}

\item{pawns}{String of pawns layout}

\item{...}{Should be left empty.}

\item{n_players}{Number of players}

\item{cfg2}{A string of a piecepack expansion (or perhaps \code{"piecepack"} for a second piecepack)}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{tibble} data frames of setups for 50 games playable with a piecepack.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   Alien City \tab \url{https://www.ludism.org/ppwiki/AlienCity} \cr
   Black Pawn Trucking \tab \url{https://www.ludism.org/ppwiki/BlackPawnTrucking} \cr
   Brain Burn \tab \url{https://www.ludism.org/ppwiki/BrainBurn} \cr
   Burbuja \tab \url{https://www.ludism.org/ppwiki/Burbuja} \cr
   Cardinal's Guards \tab \url{https://ludism.org/ppwiki/CardinalsGuards} \cr
   Cell Management \tab \url{https://www.ludism.org/ppwiki/CellManagement} \cr
   Chariots \tab \url{https://www.ludism.org/ppwiki/Chariots} \cr
   Chinese Checkers AKA Piecepack Halma \tab \url{https://www.ludism.org/ppwiki/ChineseCheckers} \cr
   Climbing Man \tab \url{https://www.ludism.org/ppwiki/ClimbingMan} \cr
   Coin Collectors \tab \url{https://www.ludism.org/ppwiki/CoinCollectors} \cr
   Crocodile Hop \tab \url{https://www.ludism.org/ppwiki/CrocodileHop} \cr
   Desfases \tab \url{https://www.ludism.org/ppwiki/Desfases} \cr
   Easy Slider \tab \url{https://www.ludism.org/ppwiki/EasySlider} \cr
   Everest \tab \url{https://www.ludism.org/ppwiki/Everest} \cr
   Four Blind Mice \tab \url{https://www.ludism.org/ppwiki/FourBlindMice} \cr
   Froggy Bottom \tab \url{https://www.ludism.org/ppwiki/FroggyBottom} \cr
   Fuji-san \tab \url{https://www.ludism.org/ppwiki/Fuji-san} \cr
   Galaxy Express \tab \url{https://www.ludism.org/ppwiki/GalaxyExpress} \cr
   Ice Floe \tab \url{https://www.ludism.org/ppwiki/IceFloe} \cr
   Iceberg \tab \url{https://www.ludism.org/ppwiki/Iceberg} \cr
   Japan \tab \url{https://www.ludism.org/ppwiki/Japan} \cr
   Lab Rats \tab \url{https://www.ludism.org/ppwiki/LabRats} \cr
   Landlocked \tab \url{https://www.ludism.org/ppwiki/Landlocked} \cr
   Ley Lines \tab \url{https://www.ludism.org/ppwiki/LeyLines} \cr
   Mathrix \tab \url{https://www.ludism.org/ppwiki/Mathrix} \cr
   One Man Thrag! \tab \url{https://www.ludism.org/ppwiki/OneManThrag} \cr
   Pass the Food \tab \url{https://www.ludism.org/ppwiki/PassTheFood} \cr
   Pawns Crossing \tab \url{https://ludism.org/ppwiki/PawnsCrossing} \cr
   Piece Gaps \tab \url{https://www.ludism.org/ppwiki/PieceGaps} \cr
   Piece Packing Pirates \tab \url{https://www.ludism.org/ppwiki/PiecePackingPirates} \cr
   Piecepack Klondike \tab \url{https://ludism.org/ppwiki/PiecepackKlondike} \cr
   Piecepackman \tab \url{https://www.ludism.org/ppwiki/Piecepackman} \cr
   Plans of Action \tab \url{https://www.ludism.org/ppwiki/PlansOfAction} \cr
   Relativity \tab \url{https://www.ludism.org/ppwiki/Relativity} \cr
   San Andreas \tab \url{https://www.ludism.org/ppwiki/SanAndreas} \cr
   Sarcophagus \tab \url{https://www.ludism.org/ppwiki/Sarcophagus} \cr
   Shopping Mall \tab \url{https://www.ludism.org/ppwiki/ShoppingMall} \cr
   Skyscrapers \tab \url{https://www.ludism.org/ppwiki/Skyscrapers} \cr
   Slides of Action \tab \url{https://www.ludism.org/ppwiki/SlidesOfAction} \cr
   Speedy Towers \tab \url{https://ludism.org/ppwiki/SpeedyTowers} \cr
   Steppin' Stones \tab \url{https://ludism.org/ppwiki/Steppin'_Stones} \cr
   The "In" Crowd \tab \url{https://www.ludism.org/ppwiki/TheInCrowd} \cr
   The Magic Bag \tab \url{https://www.ludism.org/ppwiki/The_Magic_Bag} \cr
   The Penguin Game \tab \url{https://www.ludism.org/ppwiki/PenguinGame} \cr
   Tower of Babel AKA Piecepack Accordion \tab \url{https://www.ludism.org/ppwiki/TowerOfBabel} \cr
   Tracers \tab \url{https://www.ludism.org/ppwiki/Tracers} \cr
   Triactor \tab \url{https://www.ludism.org/ppwiki/Triactor} \cr
   Tula \tab \url{https://www.ludism.org/ppwiki/Tula} \cr
   Twelve Men's Morris \tab \url{https://en.wikipedia.org/wiki/Morabaraba} \cr
   Wormholes \tab \url{https://www.ludism.org/ppwiki/Wormholes} \cr
}
}
