% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecepack_games_traditional.r
\name{piecepack_games_traditional}
\alias{piecepack_games_traditional}
\alias{piecepack_alquerque}
\alias{piecepack_awithlaknannai_mosona}
\alias{piecepack_backgammon}
\alias{piecepack_brandubh}
\alias{piecepack_cribbage}
\alias{piecepack_cribbage_board}
\alias{piecepack_four_field_kono}
\alias{piecepack_julgonu}
\alias{piecepack_ludo}
\alias{piecepack_nine_mens_morris}
\alias{piecepack_salta}
\alias{piecepack_tablut}
\alias{piecepack_twelve_mens_morris}
\title{Setups for other traditional games playable with a piecepack}
\usage{
piecepack_alquerque(has_matchsticks = FALSE)

piecepack_awithlaknannai_mosona(has_matchsticks = FALSE)

piecepack_backgammon()

piecepack_brandubh()

piecepack_cribbage()

piecepack_cribbage_board()

piecepack_four_field_kono()

piecepack_julgonu()

piecepack_ludo()

piecepack_nine_mens_morris(has_matchsticks = FALSE)

piecepack_salta(has_subpack = FALSE)

piecepack_tablut(die_width = 0.63)

piecepack_twelve_mens_morris(has_matchsticks = FALSE)
}
\arguments{
\item{has_matchsticks}{Has matchsticks}

\item{has_subpack}{Has a piecepack subpack}

\item{die_width}{Width of dice}
}
\description{
\code{tibble} data frames of setups for 11 other traditional games playable with a piecepack.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various other traditional games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   Alquerque AKA Qirkat \tab \url{https://en.wikipedia.org/wiki/Alquerque} \cr
   Awithlaknannai Mosona \tab \url{https://en.wikipedia.org/wiki/Awithlaknannai_Mosona} \cr
   Backgammon \tab \url{https://www.ludism.org/ppwiki/Backgammon} \cr
   Brandubh AKA Brandub \tab \url{http://www.cyningstan.com/game/125/brandub} \cr
   Cribbage \tab \url{https://www.ludism.org/ppwiki/Cribbage} \cr
   Four Field Kono \tab \url{https://www.ludism.org/ppwiki/FourFieldKono} \cr
   Jul-Gonu \tab \url{https://www.ludism.org/ppwiki/JulGonu} \cr
   Ludo \tab \url{https://en.wikipedia.org/wiki/Ludo_(board_game)} \cr
   Nine Men's Morris \tab \url{https://en.wikipedia.org/wiki/Nine_men\\\%27s_morris} \cr
   Salta \tab \url{https://en.wikipedia.org/wiki/Salta_(game)} \cr
   Tablut \tab \url{https://www.ludism.org/ppwiki/Tablut} \cr
}
}
