% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecepack_preview.R
\name{piecepack_preview}
\alias{piecepack_preview}
\title{Preview layout of a piecepack set}
\usage{
piecepack_preview(cfg = "piecepack")
}
\arguments{
\item{cfg}{"piecepack" or perhaps "playing_cards_expansion", "dual_piecepacks_expansion", or "subpack".}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{piecepack_preview()} returns a data frame representing a preview layout of
the pieces in a standard piecepack set.
}
\examples{
df <- piecepack_preview()
if (requireNamespace("ppcli", quietly = TRUE) &&
    packageVersion("ppcli") >= "0.3.0-2") {
  ppcli::cat_piece(df, annotate = TRUE)
}
}
