% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdf-defunct.R
\name{ppdf-defunct}
\alias{ppdf-defunct}
\alias{chess_angle}
\alias{domino_angle}
\alias{piecepack_angle}
\alias{tarot_angle}
\alias{alquerque_suit}
\alias{checker_suit}
\alias{chess_suit}
\alias{go_suit}
\alias{marble_suit}
\alias{morris_suit}
\alias{piecepack_suit}
\alias{domino_rank}
\alias{piecepack_rank}
\alias{checkers_american_checkers}
\alias{checkers_american_pool_checkers}
\alias{checkers_board}
\alias{checkers_brazilian_checkers}
\alias{checkers_breakthrough}
\alias{checkers_by_name}
\alias{checkers_canadian_checkers}
\alias{checkers_crossings}
\alias{checkers_czech_checkers}
\alias{checkers_checkers}
\alias{checkers_dameo}
\alias{checkers_english_checkers}
\alias{checkers_focus}
\alias{checkers_four_field_kono}
\alias{checkers_frisian_checkers}
\alias{checkers_gothic_checkers}
\alias{checkers_grasshopper}
\alias{checkers_international_checkers}
\alias{checkers_italian_checkers}
\alias{checkers_jamaican_checkers}
\alias{checkers_julgonu}
\alias{checkers_lines_of_action}
\alias{checkers_none}
\alias{checkers_portuguese_checkers}
\alias{checkers_russian_checkers}
\alias{checkers_spanish_checkers}
\alias{checkers_thai_checkers}
\alias{checkers_turkish_checkers}
\alias{chess_by_name}
\alias{dominoes_by_name}
\alias{dominoes_concentration}
\alias{dominoes_domino_finder}
\alias{dominoes_domino_runners}
\alias{dominoes_fujisan}
\alias{dominoes_luzon}
\alias{dominoes_none}
\alias{dominoes_patience}
\alias{dominoes_the_jubilee}
\alias{dominoes_tiles}
\alias{games_checkers}
\alias{games_chess}
\alias{games_dominoes}
\alias{games_piecepack}
\alias{games_stackpack}
\alias{piecepack_by_name}
\alias{piecepack_piecepack_accordion}
\alias{piecepack_piecepack_halma}
\alias{piecepack_piecepack_klondike}
\alias{piecepack_rect_board_tiles}
\alias{stackpack_by_name}
\title{Defunct functions}
\usage{
chess_angle(...)

domino_angle(...)

piecepack_angle(...)

tarot_angle(...)

alquerque_suit(...)

checker_suit(...)

chess_suit(...)

go_suit(...)

marble_suit(...)

morris_suit(...)

piecepack_suit(...)

domino_rank(...)

piecepack_rank(...)

checkers_american_checkers(...)

checkers_american_pool_checkers(...)

checkers_board(...)

checkers_brazilian_checkers(...)

checkers_breakthrough(...)

checkers_by_name(...)

checkers_canadian_checkers(...)

checkers_crossings(...)

checkers_czech_checkers(...)

checkers_checkers(...)

checkers_dameo(...)

checkers_english_checkers(...)

checkers_focus(...)

checkers_four_field_kono(...)

checkers_frisian_checkers(...)

checkers_gothic_checkers(...)

checkers_grasshopper(...)

checkers_international_checkers(...)

checkers_italian_checkers(...)

checkers_jamaican_checkers(...)

checkers_julgonu(...)

checkers_lines_of_action(...)

checkers_none(...)

checkers_portuguese_checkers(...)

checkers_russian_checkers(...)

checkers_spanish_checkers(...)

checkers_thai_checkers(...)

checkers_turkish_checkers(...)

chess_by_name(...)

dominoes_by_name(...)

dominoes_concentration(...)

dominoes_domino_finder(...)

dominoes_domino_runners(...)

dominoes_fujisan(...)

dominoes_luzon(...)

dominoes_none(...)

dominoes_patience(...)

dominoes_the_jubilee(...)

dominoes_tiles(...)

games_checkers(...)

games_chess(...)

games_dominoes(...)

games_piecepack(...)

games_stackpack(...)

piecepack_by_name(...)

piecepack_piecepack_accordion(...)

piecepack_piecepack_halma(...)

piecepack_piecepack_klondike(...)

piecepack_rect_board_tiles(...)

stackpack_by_name(...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
These functions are defunct in this release of ppdf.
Use the replacement function shown in the table below.
}
\details{
\tabular{ll}{
   \strong{Defunct function} \tab \strong{Replacement function} \cr
   \code{checkers_by_name()} \tab \code{checker_setup_by_name()} \cr
   \code{chess_by_name()} \tab \code{chess_setup_by_name()} \cr
   \code{dominoes_by_name()} \tab \code{domino_setup_by_name()} \cr
   \code{dominoes_concentration()} \tab \code{domino_concentration()} \cr
   \code{dominoes_domino_finder()} \tab \code{domino_finder()} \cr
   \code{dominoes_domino_runners()} \tab \code{domino_runners()} \cr
   \code{dominoes_fujisan()} \tab \code{domino_fujisan()} \cr
   \code{dominoes_luzon()} \tab \code{domino_luzon()} \cr
   \code{dominoes_none()} \tab \code{domino_none()} \cr
   \code{dominoes_patience()} \tab \code{domino_patience()} \cr
   \code{dominoes_the_jubilee()} \tab \code{domino_the_jubilee()} \cr
   \code{dominoes_tiles()} \tab \code{domino_tiles()} \cr
   \code{games_checkers()} \tab \code{checker_games()} \cr
   \code{games_chess()} \tab \code{chess_games()} \cr
   \code{games_dominoes()} \tab \code{domino_games()} \cr
   \code{games_piecepack()} \tab \code{piecepack_games()} \cr
   \code{games_stackpack()} \tab \code{stackpack_games()} \cr
   \code{piecepack_by_name()} \tab \code{piecepack_setup_by_name()} \cr
   \code{piecepack_piecepack_accordion()} \tab \code{piecepack_accordion()} \cr
   \code{piecepack_piecepack_halma()} \tab \code{piecepack_halma()} \cr
   \code{piecepack_piecepack_klondike()} \tab \code{piecepack_klondike()} \cr
   \code{piecepack_rect_board_tiles()} \tab \code{piecepack_rectangular_board()} \cr
   \code{stackpack_by_name()} \tab \code{stackpack_setup_by_name()} \cr
}
}
