% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranks.r
\name{rank_helper}
\alias{rank_helper}
\alias{chess_rank}
\alias{domino_rank}
\alias{piecepack_rank}
\alias{tarot_rank}
\title{Cast to rank integers}
\usage{
chess_rank(rank)

domino_rank(rank)

piecepack_rank(rank)

tarot_rank(rank)
}
\arguments{
\item{rank}{A numeric or character vector of rank values.}
}
\value{
An integer vector of rank values.
}
\description{
These functions help make sure rank values are the
right integer values expected by the configurations used by
\code{\link[piecepackr:game_systems]{piecepackr::game_systems()}}  and \code{\link[ppn:read_ppn]{ppn::read_ppn()}}.
If the input is numeric simply use \code{\link[=as.integer]{as.integer()}} and if the
input is character try to cast to a reasonable value.
}
\examples{
chess_rank(c("P", "N", "B", "R", "Q", "K"))
chess_rank(c("\u265f", "\u265e", "\u265d", "\u265c", "\u265b", "\u265a"))
piecepack_rank(c("n", "a", "2", "3", "4", "5"))
domino_rank(c("\U0001f06e", "\U0001f038"))
domino_rank(c("zero", "one", "eighteen"))
tarot_rank(c("ace", "ten", "jack", "queen", "knight", "king"))
}
