% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reversi.r
\name{reversi_games_variant}
\alias{reversi_games_variant}
\alias{reversi_ming_mang}
\alias{reversi_reversi}
\title{Setups for reversi variants}
\usage{
reversi_ming_mang(nrows = 8L, ncols = nrows)

reversi_reversi(nrows = 8L, ncols = nrows)
}
\arguments{
\item{nrows}{Number of rows in game board}

\item{ncols}{Number of columns in game board}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{tibble} data frames of setups for 2 reversi variants.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   Ming Mang \tab \url{https://boardgamegeek.com/boardgame/40573/ming-mang} \cr
   Reversi \tab \url{https://en.wikipedia.org/wiki/Reversi} \cr
}
}
