% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reversi.r
\name{reversi_pieces}
\alias{reversi_pieces}
\alias{reversi_board}
\alias{reversi_bits}
\title{Generate reversi boards and pieces}
\usage{
reversi_board(
  nrows = 8L,
  ncols = nrows,
  x0 = 1,
  y0 = 1,
  ...,
  side = "face",
  piece_side = paste0("board_", side),
  suit = "black",
  angle = 0
)

reversi_bits(
  ...,
  side = "face",
  piece_side = paste0("bit_", side),
  suit = 1:6,
  x = as.double(1:6),
  y = 1,
  angle = 0,
  length.out = NA_integer_
)
}
\arguments{
\item{nrows}{Number of rows in game board}

\item{ncols}{Number of columns in game board}

\item{x0}{X coordinate for the center of the first point}

\item{y0}{Y coordinate for the center of the first point}

\item{...}{Should be left empty.}

\item{side}{Either "face" or "back"}

\item{piece_side}{If \code{"board_face"} a lined board.
If \code{"board_back"} a checkered board.}

\item{suit}{Suit value (color) of board/bit.
\code{1L} is "red", \code{2L} is "black", \code{3L} is "green", \code{4L} is "blue", \code{5L} is "yellow", and \code{6L} is "white".  Will be coerced by \code{\link[=piece_suit]{piece_suit()}}.}

\item{angle}{Angle of board in degrees.}

\item{x, y}{Cartesian coordinates (numeric vectors)}

\item{length.out}{The number of pieces.
Not needed if all the arguments are the same length (or of length one)
and this length is the same as the number of desired pieces.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{reversi_board()} returns various reversi boards.
\code{reversi_bits()} returns reversi pieces.
}
