% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/none.r
\name{setup_none}
\alias{setup_none}
\alias{alquerque_none}
\alias{checker_none}
\alias{chess_none}
\alias{domino_none}
\alias{go_none}
\alias{marble_none}
\alias{piecepack_none}
\alias{stackpack_none}
\title{Zero-row data frames}
\usage{
alquerque_none()

checker_none()

chess_none()

domino_none()

go_none()

marble_none()

piecepack_none()

stackpack_none()
}
\value{
A data frame with zero rows.
}
\description{
\code{alquerque_none()}, \code{checker_none()}, \code{chess_none()}, \code{domino_none()},
\code{go_none()}, \code{marble_none()}, \code{piecepack_none()}, and \code{stackpack_none()} return zero-row data frames.
}
\examples{
checker_none()
}
