% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackpack_games.r
\name{stackpack_games}
\alias{stackpack_games}
\alias{stackpack_alice_chess}
\alias{stackpack_baroque_chess}
\alias{stackpack_chaturaji}
\alias{stackpack_chess}
\alias{stackpack_chinese_chess}
\alias{stackpack_four_seasons_chess}
\alias{stackpack_international_chess}
\alias{stackpack_japanese_chess}
\alias{stackpack_salta}
\alias{stackpack_shogi}
\alias{stackpack_ultima}
\alias{stackpack_xiangqi}
\title{Setups for games playable with a piecepack stackpack}
\usage{
stackpack_alice_chess(max_tiles = 24L)

stackpack_baroque_chess()

stackpack_chaturaji()

stackpack_chess()

stackpack_chinese_chess()

stackpack_four_seasons_chess()

stackpack_international_chess()

stackpack_japanese_chess()

stackpack_salta()

stackpack_shogi()

stackpack_ultima()

stackpack_xiangqi()
}
\arguments{
\item{max_tiles}{Maximum number of (piecepack) tiles available to build boards}
}
\description{
\code{tibble} data frames of setups for 8 other games playable with a piecepack stackpack.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   Alice Chess \tab \url{https://en.wikipedia.org/wiki/Alice_chess} \cr
   Chaturaji \tab \url{https://www.ludism.org/ppwiki/Chaturaji} \cr
   Four Seasons Chess \tab \url{https://www.chessvariants.com/historic.dir/4seiz.html} \cr
   (International) Chess \tab \url{https://www.ludism.org/ppwiki/Chess} \cr
   Salta \tab \url{https://en.wikipedia.org/wiki/Salta_(game)} \cr
   Shogi aka Japanese Chess \tab \url{https://www.ludism.org/ppwiki/Shogi} \cr
   Ultima aka Baroque Chess \tab \url{https://en.wikipedia.org/wiki/Baroque_chess} \cr
   Xiangqi AKA Chinese Chess \tab \url{https://www.ludism.org/ppwiki/Xiangqi} \cr
}
}
