% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suits.r
\name{suit_helper}
\alias{suit_helper}
\alias{alquerque_suit}
\alias{checker_suit}
\alias{chess_suit}
\alias{go_suit}
\alias{marble_suit}
\alias{piecepack_suit}
\title{Cast to suit integers}
\usage{
alquerque_suit(suit)

checker_suit(suit)

chess_suit(suit)

go_suit(suit)

marble_suit(suit)

piecepack_suit(suit)
}
\arguments{
\item{suit}{A numeric or character vector of suit values.}
}
\value{
An integer vector of suit values.
}
\description{
These functions help make sure suit values are the
right integer values expected by the configurations used by
\code{\link[piecepackr:game_systems]{piecepackr::game_systems()}}  and \code{\link[ppn:read_ppn]{ppn::read_ppn()}}.
If the input is numeric simply use \code{\link[=as.integer]{as.integer()}} and if the
input is character try to cast to a reasonable value.
}
\examples{
go_suit(c("R", "K", "G", "B", "Y", "W"))
go_suit(c("red", "black", "green", "blue", "yellow", "white"))
piecepack_suit(c("S", "M", "C", "A"))
piecepack_suit(c("suns", "moons", "crowns", "arms"))
}
