% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarot.r
\name{tarot_pieces}
\alias{tarot_pieces}
\alias{tarot_cards}
\title{Generate tarot cards}
\usage{
tarot_cards(
  ...,
  side = "face",
  piece_side = paste0("card_", side),
  suit = rep(1:4, each = 13L),
  rank = rep.int(c(1:11, 13:14), 4L),
  x = 2.5 * rep.int(1:13, 4L),
  y = 3.5 * rep(4:1, each = 13L),
  angle = 0,
  length.out = NA_integer_
)
}
\arguments{
\item{...}{Should be left empty.}

\item{side}{Either "face" or "back".}

\item{piece_side}{Either "tile_face" or "tile_back".}

\item{suit}{Suit value of cards.
\code{1L} is "hearts", \code{2L} is "spades", \code{3L} is "clubs", \code{4L} is "diamonds", and \code{5L} is "trumps".
Will be coerced by \code{\link[=tarot_suit]{tarot_suit()}}.}

\item{rank}{Rank value of cards.
Normal suits are from \code{1L} (ace) to \code{14L} (king) plus a \code{15L} joker.
Trump suits are from \code{1L} to \code{21L} plus a \code{22L} fool/excuse.
Will be coerced by \code{\link[=tarot_rank]{tarot_rank()}}.}

\item{x, y}{Cartesian coordinates (numeric vectors)}

\item{angle}{Rotation of piece (numeric vector of degrees, counter-clockwise).
Will be coerced by \code{\link[=piece_angle]{piece_angle()}}.}

\item{length.out}{The number of pieces.
Not needed if all the arguments are the same length (or of length one)
and this length is the same as the number of desired pieces.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{tarot_cards()} returns tarot cards.
}
\examples{
df <- tarot_cards(suit = "clubs", rank = "nine", x = 1.5, y = 2)
if (require("piecepackr", quietly = TRUE) &&
    packageVersion("piecepackr") >= "1.15.0-1" &&
    piecepackr:::device_supports_unicode()) {
  grid::grid.newpage()
  envir = game_systems()
  pmap_piece(df, envir = envir, default.units = "in")
}
}
