% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_games_variant.r
\name{chess_games_variant}
\alias{chess_games_variant}
\alias{chess_chess}
\alias{chess_fischer_random_chess}
\alias{chess_international_chess}
\title{Setups for chess variants}
\usage{
chess_chess(cell_width = 1)

chess_fischer_random_chess(seed = NULL, cell_width = 1)

chess_international_chess(cell_width = 1)
}
\arguments{
\item{cell_width}{Width of board cell.  Most renderers support \code{1} or \code{2}.}

\item{seed}{Seed that determines setup, either an integer or \code{NULL}}
}
\description{
\code{tibble} data frames of setups for 2 chess variants.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   (International) Chess \tab \url{https://en.wikipedia.org/wiki/Chess} \cr
   Fischer Random Chess \tab \url{https://www.chessvariants.com/diffsetup.dir/fischer.html} \cr
}
}
