% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_boards.r
\name{chess_pieces}
\alias{chess_pieces}
\alias{chess_board}
\alias{chess_bits}
\title{Generate chess boards and pieces}
\usage{
chess_board(
  nrows = 8L,
  ncols = nrows,
  x0 = 1,
  y0 = 1,
  ...,
  side = "face",
  piece_side = paste0("board_", side),
  suit = "green",
  angle = 0,
  cell_width = 1
)

chess_bits(
  ...,
  suit = rep(1:6, each = 6L),
  rank = rep(1:6, 6L),
  x = rep(1:6, 6L),
  y = rep(1:6, each = 6L),
  angle = 0,
  cell_width = 1,
  length.out = NA_integer_
)
}
\arguments{
\item{nrows}{Number of rows in game board}

\item{ncols}{Number of columns in game board}

\item{x0}{X coordinate for the center of the first cell}

\item{y0}{Y coordinate for the center of the first cell}

\item{...}{Should be left empty.}

\item{side}{Either "face" or "back"}

\item{piece_side}{If \code{"board_face"} a checkered board (as in American Checkers).
If \code{"board_back"} a lined board (as in Turkish Checkers).}

\item{suit}{Suit value (color) of board and pieces.
\code{1L} is "red", \code{2L} is "black", \code{3L} is "green", \code{4L} is "blue", \code{5L} is "yellow", and \code{6L} is "white".
Will be coerced by \code{\link[=chess_suit]{chess_suit()}}.}

\item{angle}{Rotation of bit/board (numeric vector of degrees, counter-clockwise).
Will be coerced by \code{\link[=chess_angle]{chess_angle()}}.}

\item{cell_width}{Width of board cell.
\code{\link[piecepackr:game_systems]{piecepackr::game_systems()}} supports \code{1} or \code{2}.}

\item{rank}{Rank value of chess pieces from \code{1L} (pawn) to \code{6L} (king).
Will be coerced by \code{\link[=chess_rank]{chess_rank()}}.}

\item{x, y}{Cartesian coordinates (numeric vectors)}

\item{length.out}{The number of pieces.
Not needed if all the arguments are the same length (or of length one)
and this length is the same as the number of desired pieces.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{chess_board()} returns a square checkered board of desired size and type.
\code{chess_bits()} returns chess pieces.
}
