% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_info.r
\name{games_info}
\alias{games_info}
\alias{chess_games}
\alias{checkers_games}
\alias{domino_games}
\alias{piecepack_games}
\alias{stackpack_games}
\title{Data frames with game info}
\usage{
chess_games()

checkers_games()

domino_games()

piecepack_games()

stackpack_games()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with character columns "game", "methods", "comment", and "url"
for game name, setup function name(s), possible comment, and url for more information.
}
\description{
\code{chess_games()}, \code{domino_games()}, \code{checkers_games()}, \code{piecepack_games()}, and \code{stackpack_games()}
contain information about the games
whose setups are provided by this package.
}
\examples{
head(piecepack_games())
# Number of games for each game system
nrow(chess_games())
nrow(checkers_games())
nrow(domino_games())
nrow(piecepack_games())
nrow(stackpack_games())
}
