% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_info.r
\name{games_info}
\alias{games_info}
\alias{alquerque_games}
\alias{chess_games}
\alias{checker_games}
\alias{domino_games}
\alias{go_games}
\alias{marble_games}
\alias{morris_games}
\alias{piecepack_games}
\alias{reversi_games}
\alias{stackpack_games}
\alias{tarot_games}
\title{Data frames with game info}
\usage{
alquerque_games()

chess_games()

checker_games()

domino_games()

go_games()

marble_games()

morris_games()

piecepack_games()

reversi_games()

stackpack_games()

tarot_games()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with character columns "game", "methods", "comment", and "url"
for game name, setup function name(s), possible comment, and url for more information.
}
\description{
\code{chess_games()}, \code{domino_games()}, \code{checker_games()}, \code{piecepack_games()}, and \code{stackpack_games()}
contain information about the games
whose setups are provided by this package.
}
\examples{
head(piecepack_games())
# Number of games for each game system
nrow(alquerque_games())
nrow(chess_games())
nrow(checker_games())
nrow(domino_games())
nrow(go_games())
nrow(marble_games())
nrow(morris_games())
nrow(piecepack_games())
nrow(stackpack_games())
}
