% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecepack_boards.r
\name{piecepack_board}
\alias{piecepack_board}
\alias{piecepack_donut_board}
\alias{piecepack_rectangular_board}
\title{Setup various game boards using piecepack tiles}
\usage{
piecepack_donut_board(
  ...,
  side = "back",
  piece_side = paste0("tile_", side),
  suit = rep(1:4, each = 6L),
  rank = rep.int(1:6, 4L),
  cfg = "piecepack",
  x0 = 1,
  y0 = 1,
  angle = 0
)

piecepack_rectangular_board(
  nrows = 8L,
  ncols = 8L,
  x0 = 1,
  y0 = 1,
  ...,
  max_tiles = 24L,
  suit = rep.int(1:4, 6L),
  rank = rep(1:6, each = 4L),
  cfg = "piecepack",
  angle = 0
)
}
\arguments{
\item{...}{Should be left empty.}

\item{side}{Either "face" or "back".}

\item{piece_side}{Either "tile_face" or "tile_back".}

\item{suit}{Vector of suit values to use for tile back (will be repeated and coerced by \code{\link[=piecepack_suit]{piecepack_suit()}}).}

\item{rank}{Vector of rank values to use for tile back (will be repeated and coerced by \code{\link[=piecepack_rank]{piecepack_rank()}}).}

\item{cfg}{"piecepack" or perhaps "playing_cards_expansion", "dual_piecepacks_expansion", or "subpack".}

\item{x0}{X coordinate for the center of the first cell/point for \code{piecepack_rectangular_board()}.
X coordinate for center of bottom-left tile for \code{piecepack_donut_board()}.}

\item{y0}{Y coordinate for the center of the first cell/point for \code{piecepack_rectangular_board()}.
Y coordinate for center of bottom-left tile for \code{piecepack_donut_board()}.}

\item{angle}{Rotation of piece (numeric vector of degrees, counter-clockwise).
Will be coerced by \code{\link[=piecepack_angle]{piecepack_angle()}}.}

\item{nrows}{Number of rows in game board}

\item{ncols}{Number of columns in game board}

\item{max_tiles}{Maximum number of tiles that can be used}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{piecepack_rectangular_board()} returns a tibble data frame  of a rectangular board of desired size
made using a maximum number of piecepack tiles.
It will use either \dQuote{cells}, \dQuote{points}, and/or \dQuote{rivers} as necessary.
\code{piecepack_donut_board()} returns 24 tiles arranged 5x5 with hole in middle.
}
