% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecepack_games_other.r
\name{piecepack_games_other}
\alias{piecepack_games_other}
\alias{piecepack_12345ive}
\alias{piecepack_breakthrough}
\alias{piecepack_change_change}
\alias{piecepack_crossings}
\alias{piecepack_dao}
\alias{piecepack_evade}
\alias{piecepack_grasshopper}
\alias{piecepack_lines_of_action}
\alias{piecepack_lukawan}
\alias{piecepack_quatri}
\title{Setups for other modern games playable with a piecepack}
\usage{
piecepack_12345ive()

piecepack_breakthrough()

piecepack_change_change(seed = NULL, coins = NULL)

piecepack_crossings()

piecepack_dao()

piecepack_evade()

piecepack_grasshopper()

piecepack_lines_of_action()

piecepack_lukawan()

piecepack_quatri()
}
\arguments{
\item{seed}{Seed that determines setup, either an integer or \code{NULL}}

\item{coins}{String of coin layout}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the following columns:
\itemize{
\item "piece_side"
\item "suit"
\item "rank"
\item "cfg" (optional, defaults to \code{"piecepack"})
\item "x"
\item "y"
\item "angle" (optional, defaults to \code{0}).
}

This data frame is compatible with \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}} and likely \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.
}
\description{
\code{tibble} data frames of setups for 10 other games playable with a piecepack.
Data frame output can usually be plotted with \code{pmap_piece(df, default.units = "in")}.
}
\details{
Here are links for more information about the various other games:\tabular{ll}{
   \strong{Game} \tab \strong{URL} \cr
   12345ive! \tab \url{https://boardgamegeek.com/boardgame/154644/12345ive} \cr
   Breakthrough \tab \url{https://en.wikipedia.org/wiki/Breakthrough_(board_game)} \cr
   Crossings \tab \url{https://en.wikipedia.org/wiki/Crossings_(game)} \cr
   Change Change \tab \url{https://www.ludism.org/ppwiki/ChangeChange} \cr
   Dao \tab \url{https://boardgamegeek.com/boardgame/948/dao} \cr
   Grasshopper \tab \url{http://www.cyningstan.com/game/71/grasshopper} \cr
   Evade \tab \url{https://www.ludism.org/ppwiki/Evade} \cr
   Lines of Action \tab \url{https://en.wikipedia.org/wiki/Lines_of_Action} \cr
   Lukawan \tab \url{https://ludism.org/ppwiki/Lukawan} \cr
   Quatri \tab \url{https://www.ludism.org/ppwiki/Quatri} \cr
}
}
