% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdf-deprecated.r
\name{ppdf-deprecated}
\alias{ppdf-deprecated}
\alias{piecepack_rect_board_tiles}
\title{Deprecated functions}
\usage{
piecepack_rect_board_tiles(
  nrows = 8L,
  ncols = 8L,
  x0 = 1,
  y0 = 1,
  max_tiles = 24L,
  suit = rep.int(1:4, 6L),
  rank = rep(1:6, each = 4L)
)
}
\arguments{
\item{nrows, ncols, x0, y0, max_tiles, suit, rank}{See \code{\link[=piecepack_rectangular_board]{piecepack_rectangular_board()}}.}
}
\description{
These functions are Deprecated in this release of ppdf,
they will be marked as Defunct and removed in a future version.
}
