% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdf-deprecated.r
\name{ppdf-deprecated}
\alias{ppdf-deprecated}
\alias{checkers_american_checkers}
\alias{checkers_american_pool_checkers}
\alias{checkers_board}
\alias{checkers_brazilian_checkers}
\alias{checkers_breakthrough}
\alias{checkers_by_name}
\alias{checkers_canadian_checkers}
\alias{checkers_crossings}
\alias{checkers_czech_checkers}
\alias{checkers_checkers}
\alias{checkers_dameo}
\alias{checkers_english_checkers}
\alias{checkers_focus}
\alias{checkers_four_field_kono}
\alias{checkers_frisian_checkers}
\alias{checkers_gothic_checkers}
\alias{checkers_grasshopper}
\alias{checkers_international_checkers}
\alias{checkers_italian_checkers}
\alias{checkers_jamaican_checkers}
\alias{checkers_julgonu}
\alias{checkers_lines_of_action}
\alias{checkers_none}
\alias{checkers_portuguese_checkers}
\alias{checkers_russian_checkers}
\alias{checkers_spanish_checkers}
\alias{checkers_thai_checkers}
\alias{checkers_turkish_checkers}
\alias{chess_by_name}
\alias{dominoes_by_name}
\alias{dominoes_concentration}
\alias{dominoes_domino_finder}
\alias{dominoes_domino_runners}
\alias{dominoes_fujisan}
\alias{dominoes_luzon}
\alias{dominoes_none}
\alias{dominoes_patience}
\alias{dominoes_the_jubilee}
\alias{dominoes_tiles}
\alias{games_checkers}
\alias{games_chess}
\alias{games_dominoes}
\alias{games_piecepack}
\alias{games_stackpack}
\alias{piecepack_by_name}
\alias{piecepack_piecepack_accordion}
\alias{piecepack_piecepack_halma}
\alias{piecepack_piecepack_klondike}
\alias{piecepack_rect_board_tiles}
\alias{stackpack_by_name}
\title{Deprecated functions}
\usage{
checkers_american_checkers(cell_width = 1)

checkers_american_pool_checkers(cell_width = 1)

checkers_board(cell_width = 1)

checkers_brazilian_checkers(cell_width = 1)

checkers_breakthrough(cell_width = 1)

checkers_by_name(
  name,
  ...,
  getter = function(x) get(x, envir = getNamespace("ppdf"))
)

checkers_canadian_checkers(cell_width = 1)

checkers_crossings(cell_width = 1)

checkers_czech_checkers(cell_width = 1)

checkers_checkers(cell_width = 1)

checkers_dameo(cell_width = 1)

checkers_english_checkers(cell_width = 1)

checkers_focus(cell_width = 1)

checkers_four_field_kono(cell_width = 1)

checkers_frisian_checkers(cell_width = 1)

checkers_gothic_checkers(cell_width = 1)

checkers_grasshopper(cell_width = 1)

checkers_international_checkers(cell_width = 1)

checkers_italian_checkers(cell_width = 1)

checkers_jamaican_checkers(cell_width = 1)

checkers_julgonu(cell_width = 1)

checkers_lines_of_action(cell_width = 1)

checkers_none()

checkers_portuguese_checkers(cell_width = 1)

checkers_russian_checkers(cell_width = 1)

checkers_spanish_checkers(cell_width = 1)

checkers_thai_checkers(cell_width = 1)

checkers_turkish_checkers(cell_width = 1)

chess_by_name(
  name,
  ...,
  getter = function(x) get(x, envir = getNamespace("ppdf"))
)

dominoes_by_name(
  name,
  ...,
  getter = function(x) get(x, envir = getNamespace("ppdf"))
)

dominoes_concentration(seed = NULL)

dominoes_domino_finder(seed = NULL)

dominoes_domino_runners(seed = NULL)

dominoes_fujisan(seed = NULL)

dominoes_luzon(seed = NULL)

dominoes_none()

dominoes_patience(seed = NULL)

dominoes_the_jubilee(seed = NULL)

dominoes_tiles(
  n = 7,
  ...,
  side = "face",
  piece_side = paste0("tile_", side),
  suit = sequence(n:1, from = 1:n),
  rank = rep.int(1:n, n:1),
  cfg = "dominoes",
  x = sequence(n:1, from = 1:n),
  y = 2 * rep.int(n:1, n:1) - 0.5,
  angle = 0
)

games_checkers()

games_chess()

games_dominoes()

games_piecepack()

games_stackpack()

piecepack_by_name(
  name,
  ...,
  getter = function(x) get(x, envir = getNamespace("ppdf"))
)

piecepack_piecepack_accordion(seed = NULL, tiles = NULL)

piecepack_piecepack_halma()

piecepack_piecepack_klondike(seed = NULL)

piecepack_rect_board_tiles(
  nrows = 8L,
  ncols = 8L,
  x0 = 1,
  y0 = 1,
  max_tiles = 24L,
  suit = rep.int(1:4, 6L),
  rank = rep(1:6, each = 4L)
)

stackpack_by_name(
  name,
  ...,
  getter = function(x) get(x, envir = getNamespace("ppdf"))
)
}
\arguments{
\item{cell_width}{Width of board cell.  Most renderers support \code{1} or \code{2}.}

\item{name}{Game name.  Will be normalized by \code{\link[=normalize_name]{normalize_name()}}.}

\item{...}{Additional arguments to pass to the underlying setup function.}

\item{getter}{Function with a single function name argument to use.
By default will just look for functions in this package but
a developer may want to consider \code{\link[=dynGet]{dynGet()}} or a
wrapper around \code{\link[=get]{get()}} with a custom \code{envir} argument.}

\item{seed}{Seed that determines setup, either an integer or \code{NULL}}

\item{n}{The number of doubles in a set e.g. \code{n = 7} for double-6 set, \code{n = 10} for a double-9 set, \code{n = 13} for a double-12 set.}

\item{side}{Either "face" or "back".}

\item{piece_side}{Either "tile_face" or "tile_back".}

\item{suit}{"Bottom" half of domino.  Will be coerced by \code{\link[=domino_suit]{domino_suit()}}.}

\item{rank}{"Top" half of domino.  Will be coerced by \code{\link[=domino_rank]{domino_rank()}}.}

\item{cfg}{"dominoes" or perhaps "dominoes_black", "dominoes_blue", "dominoes_green", "dominoes_red", "dominoes_white", or "dominoes_yellow".}

\item{x, y}{Cartesian coordinates (numeric vectors)}

\item{angle}{Rotation of dominoes (numeric vector of degrees, counter-clockwise).  Will be coerced by \code{\link[=domino_angle]{domino_angle()}}.}

\item{tiles}{String of tile layout}

\item{nrows}{Number of rows in game board}

\item{ncols}{Number of columns in game board}

\item{x0}{X coordinate for the center of the first cell/point for \code{piecepack_rectangular_board()}.
X coordinate for center of bottom-left tile for \code{piecepack_donut_board()}.}

\item{y0}{Y coordinate for the center of the first cell/point for \code{piecepack_rectangular_board()}.
Y coordinate for center of bottom-left tile for \code{piecepack_donut_board()}.}

\item{max_tiles}{Maximum number of tiles that can be used}
}
\description{
These functions are Deprecated in this release of ppdf,
they will be marked as Defunct and removed in a future version.
}
\details{
\tabular{ll}{
   \strong{Deprecated function} \tab \strong{Replacement function} \cr
   \code{checkers_by_name()} \tab \code{checkers_setup_by_name()} \cr
   \code{chess_by_name()} \tab \code{chess_setup_by_name()} \cr
   \code{dominoes_by_name()} \tab \code{domino_setup_by_name()} \cr
   \code{dominoes_concentration()} \tab \code{domino_concentration()} \cr
   \code{dominoes_domino_finder()} \tab \code{domino_finder()} \cr
   \code{dominoes_domino_runners()} \tab \code{domino_runners()} \cr
   \code{dominoes_fujisan()} \tab \code{domino_fujisan()} \cr
   \code{dominoes_luzon()} \tab \code{domino_luzon()} \cr
   \code{dominoes_none()} \tab \code{domino_none()} \cr
   \code{dominoes_patience()} \tab \code{domino_patience()} \cr
   \code{dominoes_the_jubilee()} \tab \code{domino_the_jubilee()} \cr
   \code{dominoes_tiles()} \tab \code{domino_tiles()} \cr
   \code{games_checkers()} \tab \code{checkers_games()} \cr
   \code{games_chess()} \tab \code{chess_games()} \cr
   \code{games_dominoes()} \tab \code{domino_games()} \cr
   \code{games_piecepack()} \tab \code{piecepack_games()} \cr
   \code{games_stackpack()} \tab \code{stackpack_games()} \cr
   \code{piecepack_by_name()} \tab \code{piecepack_setup_by_name()} \cr
   \code{piecepack_piecepack_accordion()} \tab \code{piecepack_accordion()} \cr
   \code{piecepack_piecepack_halma()} \tab \code{piecepack_halma()} \cr
   \code{piecepack_piecepack_klondike()} \tab \code{piecepack_klondike()} \cr
   \code{piecepack_rect_board_tiles()} \tab \code{piecepack_rectangular_board()} \cr
   \code{stackpack_by_name()} \tab \code{stackpack_setup_by_name()} \cr
}
}
