% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suits.r
\name{suit_helper}
\alias{suit_helper}
\alias{alquerque_suit}
\alias{checker_suit}
\alias{chess_suit}
\alias{domino_suit}
\alias{go_suit}
\alias{marble_suit}
\alias{morris_suit}
\alias{piecepack_suit}
\alias{tarot_suit}
\title{Cast to suit integers}
\usage{
alquerque_suit(suit)

checker_suit(suit)

chess_suit(suit)

domino_suit(suit)

go_suit(suit)

marble_suit(suit)

morris_suit(suit)

piecepack_suit(suit)

tarot_suit(suit)
}
\arguments{
\item{suit}{A numeric or character vector of suit values.}
}
\value{
An integer vector of suit values.
}
\description{
These functions help make sure suit values are the
right integer values expected by the configurations used by
\code{\link[piecepackr:game_systems]{piecepackr::game_systems()}}  and \code{\link[ppn:read_ppn]{ppn::read_ppn()}}.
If the input is numeric it simply uses \code{\link[=as.integer]{as.integer()}} and if the
input is character it tries to cast to a reasonable value.
}
\examples{
go_suit(c("red", "black", "green", "blue", "yellow", "white"))
go_suit(c("R", "K", "G", "B", "Y", "W"))
piecepack_suit(c("suns", "moons", "crowns", "arms"))
piecepack_suit(c("S", "M", "C", "A"))
domino_suit(c("\U0001f06e", "\U0001f038"))
domino_suit(c("zero", "one", "eighteen"))
tarot_suit(c("\u2665", "\u2660", "\u2663", "\u2666"))
tarot_suit(c("hearts", "spades", "clubs", "diamonds", "trumps"))
}
